+++ backend/http/auth.go
@@ -124,7 +124,7 @@ func loginHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (in
 	return printToken(w, r, d.user) // Pass the data object
 }
 
-// logoutHandler handles user logout, specifically used for OIDC.
+// logoutHandler handles user logout, specifically used for OIDC or proxy users.
 // @Summary User Logout
 // @Description logs a user out of the application.
 // @Tags Auth
@@ -136,12 +136,15 @@ func logoutHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (i
 		origin = fmt.Sprintf("%s://%s", getScheme(r), r.Host)
 	}
 	logoutUrl := fmt.Sprintf("%s/login", origin)
+	proxyRedirectUrl := config.Auth.Methods.ProxyAuth.LogoutRedirectUrl
+	oidcRedirectUrl := settings.Config.Auth.Methods.OidcAuth.LogoutRedirectUrl
+	userLoginMethod := d.user.LoginMethod
 
-	oidcCfg := settings.Config.Auth.Methods.OidcAuth
-	if oidcCfg.Enabled && oidcCfg.LogoutRedirectUrl != "" {
-		logoutUrl = oidcCfg.LogoutRedirectUrl
-		http.Redirect(w, r, logoutUrl, http.StatusFound)
-		return 302, nil
+	if d.user.LoginMethod == users.LoginMethodProxy && proxyRedirectUrl != "" {
+		logoutUrl = proxyRedirectUrl
+	}
+	if userLoginMethod == users.LoginMethodOidc && oidcRedirectUrl != "" {
+		logoutUrl = oidcRedirectUrl
 	}
 
 	http.Redirect(w, r, logoutUrl, http.StatusFound)