+++ backend/common/settings/structs.go
@@ -21,7 +21,7 @@ type Settings struct {
 }
 
 type Server struct {
-	NumImageProcessors           int         `json:"numImageProcessors"`           // number of concurrent image processing jobs
+	NumImageProcessors           int         `json:"numImageProcessors"`           // number of concurrent image processing jobs used to create previews, default is 4,
 	Socket                       string      `json:"socket"`                       // socket to listen on
 	TLSKey                       string      `json:"tlsKey"`                       // path to TLS key
 	TLSCert                      string      `json:"tlsCert"`                      // path to TLS cert
@@ -42,6 +42,7 @@ type Server struct {
 	SourceMap     map[string]Source `json:"-" validate:"omitempty"` // uses realpath as key
 	NameToSource  map[string]Source `json:"-" validate:"omitempty"` // uses name as key
 	DefaultSource Source            `json:"-" validate:"omitempty"`
+	PdfAvailable  bool              `json:"-"` // used internally if compiled with pdf support
 }
 
 type Integrations struct {