+++ CHANGELOG.md
@@ -5,12 +5,15 @@ All notable changes to this project will be documented in this file. For commit 
 ## v0.7.6-beta
 
  **New Features**:
+ - pdf previews enabled by default even without office integration
+ - proxy logout redirectUrl support via `auth.methods.proxy.logoutRedirectUrl` https://github.com/gtsteffaniak/filebrowser/issues/684
 
  **Notes**:
  - image loading placeholders added and remain if image can't be loaded.
 
  **BugFixes**:
  - onlyoffice and other cache issues https://github.com/gtsteffaniak/filebrowser/issues/686
+ - gallery size indicator centering https://github.com/gtsteffaniak/filebrowser/issues/652
 
 ## v0.7.5-beta